% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{This is the Euclidean function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{a number}
}
\value{
the GCD of two given numbers
}
\description{
This is a great function rarely seen in nature. See here: https://www.google.com/url?sa=t&rct=j&q=&esrc=s&source=web&cd=12&cad=rja&uact=8&ved=2ahUKEwj6nP7jsdHkAhVBtIsKHUfVCM8QFjALegQIABAB&url=https%3A%2F%2Fen.wikipedia.org%2Fwiki%2FEuclidean_algorithm&usg=AOvVaw2vg5R_wmq5L2pEWXWyvzB1
check input type
obtain maximum and minimum of the inputs
calculate the remainder
repeat process until remainder is equal to zero
return one to the last remainder
}
\examples{
 
euclidean(100,10)
}
