\name{testlab-package}
\alias{testlab-package}
\alias{testlab}
\docType{package}
\title{
\packageTitle{testlab}
}
\description{
\packageDescription{testlab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testlab}
\packageIndices{testlab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{testlab}

Maintainer: \packageMaintainer{testlab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
